+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated examples for `terra-slide-panel-manager`.
+
 ## 1.19.0 - (April 12, 2023)
 
 * Added
@@ -17,7 +20,7 @@
 * Changed
   * Updated docs for `terra-slide-group`.
   * Updated doc and added test for terra-collapsible-menu-view
-  
+
 * Added
   * Added package-lock.json
   * Added test for `terra-slide-panel`.